Cleaned spacing in VScode & removed U0 in toolchange
I commented out the  command(MMU_CMD_U0);
in the tool_change() since I checked the MMU code does manage whether to do an unload/not with the succeeding "command(MMU_CMD_T0 + index);" if the currently assigned extruder has changed.

so as not to sound confusing, the goal of keeping the extruder cleared with any filament in the gears is still there, I just removed an excess MMU command that is already managed by the succeeding command anyway.